+++ lib/src/features/auth/register/register_screen_vm.g.dart
@@ -24,68 +24,6 @@ mixin _$RegisterScreenVm on _RegisterScreenVm, Store {
     });
   }
 
-  final _$nameControllerAtom = Atom(name: '_RegisterScreenVm.nameController');
-
-  @override
-  TextEditingController get nameController {
-    _$nameControllerAtom.reportRead();
-    return super.nameController;
-  }
-
-  @override
-  set nameController(TextEditingController value) {
-    _$nameControllerAtom.reportWrite(value, super.nameController, () {
-      super.nameController = value;
-    });
-  }
-
-  final _$surnameControllerAtom =
-      Atom(name: '_RegisterScreenVm.surnameController');
-
-  @override
-  TextEditingController get surnameController {
-    _$surnameControllerAtom.reportRead();
-    return super.surnameController;
-  }
-
-  @override
-  set surnameController(TextEditingController value) {
-    _$surnameControllerAtom.reportWrite(value, super.surnameController, () {
-      super.surnameController = value;
-    });
-  }
-
-  final _$emailControllerAtom = Atom(name: '_RegisterScreenVm.emailController');
-
-  @override
-  TextEditingController get emailController {
-    _$emailControllerAtom.reportRead();
-    return super.emailController;
-  }
-
-  @override
-  set emailController(TextEditingController value) {
-    _$emailControllerAtom.reportWrite(value, super.emailController, () {
-      super.emailController = value;
-    });
-  }
-
-  final _$birthdayControllerAtom =
-      Atom(name: '_RegisterScreenVm.birthdayController');
-
-  @override
-  TextEditingController get birthdayController {
-    _$birthdayControllerAtom.reportRead();
-    return super.birthdayController;
-  }
-
-  @override
-  set birthdayController(TextEditingController value) {
-    _$birthdayControllerAtom.reportWrite(value, super.birthdayController, () {
-      super.birthdayController = value;
-    });
-  }
-
   final _$checkboxValueAtom = Atom(name: '_RegisterScreenVm.checkboxValue');
 
   @override
@@ -98,21 +36,6 @@ mixin _$RegisterScreenVm on _RegisterScreenVm, Store {
   set checkboxValue(bool value) {
     _$checkboxValueAtom.reportWrite(value, super.checkboxValue, () {
       super.checkboxValue = value;
-    });
-  }
-
-  final _$maskFormatterAtom = Atom(name: '_RegisterScreenVm.maskFormatter');
-
-  @override
-  MaskTextInputFormatter get maskFormatter {
-    _$maskFormatterAtom.reportRead();
-    return super.maskFormatter;
-  }
-
-  @override
-  set maskFormatter(MaskTextInputFormatter value) {
-    _$maskFormatterAtom.reportWrite(value, super.maskFormatter, () {
-      super.maskFormatter = value;
     });
   }
 
@@ -135,28 +58,6 @@ mixin _$RegisterScreenVm on _RegisterScreenVm, Store {
       ActionController(name: '_RegisterScreenVm');
 
   @override
-  String? emailValidator(String? value) {
-    final _$actionInfo = _$_RegisterScreenVmActionController.startAction(
-        name: '_RegisterScreenVm.emailValidator');
-    try {
-      return super.emailValidator(value);
-    } finally {
-      _$_RegisterScreenVmActionController.endAction(_$actionInfo);
-    }
-  }
-
-  @override
-  String? fullNameValidator(String? value) {
-    final _$actionInfo = _$_RegisterScreenVmActionController.startAction(
-        name: '_RegisterScreenVm.fullNameValidator');
-    try {
-      return super.fullNameValidator(value);
-    } finally {
-      _$_RegisterScreenVmActionController.endAction(_$actionInfo);
-    }
-  }
-
-  @override
   void onChangedCheckbox(bool? value) {
     final _$actionInfo = _$_RegisterScreenVmActionController.startAction(
         name: '_RegisterScreenVm.onChangedCheckbox');
@@ -168,11 +69,11 @@ mixin _$RegisterScreenVm on _RegisterScreenVm, Store {
   }
 
   @override
-  void registerBtnClicked() {
+  void checkIsFormEntered() {
     final _$actionInfo = _$_RegisterScreenVmActionController.startAction(
-        name: '_RegisterScreenVm.registerBtnClicked');
+        name: '_RegisterScreenVm.checkIsFormEntered');
     try {
-      return super.registerBtnClicked();
+      return super.checkIsFormEntered();
     } finally {
       _$_RegisterScreenVmActionController.endAction(_$actionInfo);
     }
@@ -182,12 +83,7 @@ mixin _$RegisterScreenVm on _RegisterScreenVm, Store {
   String toString() {
     return '''
 vmState: ${vmState},
-nameController: ${nameController},
-surnameController: ${surnameController},
-emailController: ${emailController},
-birthdayController: ${birthdayController},
 checkboxValue: ${checkboxValue},
-maskFormatter: ${maskFormatter},
 isFormEntered: ${isFormEntered}
     ''';
   }