+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - feat: CodeVm - add authorization via OTP
 - feat: SignInScreenVm - add phone authorization
 - feat: OnboardingScreen
+- feat: MainScreen + PlacesScreen
 - feat: profile screen
 
 ## 0.3.0